Changes for signed/unsigned conflicts and other type nits.  C++ is much more picky about these than C, but this is probably justified.  Sadly, some of the changes are pretty ugly.

+++ ep/ep_pcvt.c
@@ -47,7 +47,7 @@ ep_pcvt_str(
 	size_t osize,
 	const char *val)
 {
-	int vl;
+	size_t vl;
 
 	if (val == NULL)
 	{
